requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
wordcloud
Pillow
regex
emoji
demoji
